shop/requirements.txt
celery==5.2.7
django-bootstrap-modal-forms==2.2.0
django-filter==22.1
django-mptt-admin==2.3.0
django-mptt==0.13.4
django-property-filter==1.1.1
django-widget-tweaks==1.4.12
Django==4.0.7
djangorestframework==3.14.0
flower==1.2.0
gevent==22.10.2
greenlet==2.0.1
Markdown==3.4.1
Pillow==9.2.0
requests==2.28.1
PyYAML==6.0
sorl-thumbnail==12.9.0